+++ shardis-auth/src/main/java/com/shardis/auth/controllers/rest/AuthRestController.java
@@ -1,7 +1,5 @@
-package com.shardis.api.controllers.rest;
+package com.shardis.auth.controllers.rest;
 
-
-import com.shardis.dto.commons.ValueObject;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -10,19 +8,12 @@ import java.security.Principal;
 /**
  * Created by Tomasz Kucharzyk
  */
-
 @RestController
 @RequestMapping("/")
-public class TestRestController {
-
-  @RequestMapping("/title")
-  public ValueObject getTitle() {
-    return new ValueObject("API SERVER IS ALIVE");
-  }
+public class AuthRestController {
 
     @RequestMapping("/user")
     public Principal user(Principal user) {
         return user;
     }
-
 }