+++ shardis-ui/src/main/java/com/shardis/ShardisUiApplication.java
@@ -5,6 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
 import org.springframework.boot.context.embedded.ErrorPage;
 import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
@@ -13,20 +14,21 @@ import org.springframework.http.HttpStatus;
  * Created by Tomasz Kucharzyk
  */
 @SpringBootApplication
+@EnableDiscoveryClient
 @EnableCaching
 @EnableZuulProxy
 public class ShardisUiApplication {
 
-  public static void main(String[] args) {
-    SpringApplication.run(ShardisUiApplication.class, args);
-  }
+    public static void main(String[] args) {
+        SpringApplication.run(ShardisUiApplication.class, args);
+    }
 
-  @Bean
-  public EmbeddedServletContainerCustomizer containerCustomizer() {
+    @Bean
+    public EmbeddedServletContainerCustomizer containerCustomizer() {
 
-    return (container -> {
-      ErrorPage error404Page = new ErrorPage(HttpStatus.NOT_FOUND, "/");
-      container.addErrorPages(error404Page);
-    });
-  }
+        return (container -> {
+            ErrorPage error404Page = new ErrorPage(HttpStatus.NOT_FOUND, "/");
+            container.addErrorPages(error404Page);
+        });
+    }
 }