+++ shardis-ui/config/spec-bundle.js
@@ -14,8 +14,8 @@ require('zone.js');
 require('zone.js/dist/long-stack-trace-zone.js');
 require('zone.js/dist/jasmine-patch.js');
 
-require('rxjs/add/operator/map');
-require('rxjs/add/operator/mergeMap');
+require('rxjs/Rx');
+
 
 var testing = require('angular2/testing');
 var browser = require('angular2/platform/testing/browser');