+++ shardis-ui/src/main/frontend/vendor.ts
@@ -5,8 +5,7 @@ import 'angular2/http';
 import 'angular2/router';
 
 // RxJS
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mergeMap';
+import 'rxjs/Rx';
 
 if ('production' === ENV) {
   // Production