+++ shardis-ui/config/spec-bundle.js
@@ -9,10 +9,11 @@
 */
 Error.stackTraceLimit = Infinity;
 require('core-js');
+require('ts-helpers');
 
-require('zone.js');
-require('zone.js/dist/long-stack-trace-zone.js');
-require('zone.js/dist/jasmine-patch.js');
+require('zone.js/dist/zone');
+require('zone.js/dist/long-stack-trace-zone');
+require('zone.js/dist/jasmine-patch');
 
 require('rxjs/Rx');