+++ shardis-ui/src/main/frontend/polyfills.ts
@@ -1,8 +1,10 @@
-// Ccre.js 
+// Ccre.js
 import  'core-js';
 
 // Zone.js
 import 'zone.js/lib/browser/browser';
+
+import 'ts-helpers';
 
 if ('production' !== ENV) {
   Error['stackTraceLimit'] = Infinity;