+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * oauth2 security
 * Updated Angular to beta 14
 * Included all rxjs
+* updated Angular to beta 15
+* updated some dependencies
 
 ### 1.2.2 (01.04.2016)
 * Updated Angular to beta 13